refactor: DiaryDataManager -> DiaryManager 이름 변경, DiaryManager에 싱글톤 패턴 적용, CRUD 메서드에 throw 추가
