+++ Diary/Source/Protocol/DiaryManageable.swift
@@ -7,8 +7,8 @@ import CoreData
 protocol CoreDataManageable {
     associatedtype Object: ManagedObjectModel
     
-    func add(_ object: Object?)
-    func fetchObjects() -> [Object]
-    func update(_ object: Object)
-    func remove(_ object: Object)
+    func add(_ object: Object?) throws
+    func fetchObjects() throws -> [Object] 
+    func update(_ object: Object) throws
+    func remove(_ object: Object) throws
 }
